feat: add solar panel segmentation with UNet ONNX inference
- Convert solar_unet.pth (42K params) to ONNX via scripts/convert_solar_model.py
- Dequantize embeddings at load time (int8 * scale → float32) in zarr-source
- Add onnxruntime-web with Vite plugin to serve WASM/MJS files and bypass
  Vite's module transform for ORT's dynamic imports
- Create segment.ts with sliding-window inference, probability caching,
  connected-component polygonization, and threshold-adjustable re-polygonization
- Create SegmentPanel.svelte with detect button, progress bar, threshold
  slider, error display, and result count
- Wire into ToolSwitcher (Solar tab with Sun icon) and App.svelte
  (GeoJSON polygon overlay in semi-transparent orange)

+++ packages/maplibre-zarr-tessera/src/types.ts
@@ -82,7 +82,7 @@ export interface ZarrTesseraEvents {
 export interface TileEmbeddings {
   ci: number;
   cj: number;
-  emb: Int8Array;         // [h * w * nBands] raw embedding bytes
+  emb: Float32Array;      // [h * w * nBands] dequantized float32 embeddings
   scales: Float32Array;   // [h * w] scale values
   width: number;
   height: number;
